+++ modules/github-repo/main.tf
@@ -14,19 +14,22 @@ resource "github_repository" "this" {
   allow_update_branch  = var.config.repository.allowUpdateBranch
   has_issues           = var.config.repository.hasIssues
 
-  dynamic "pages" {
-    for_each = var.config.pages != null ? [var.config.pages] : []
+
+}
+
+# GitHub Pages configuration
+resource "github_repository_pages" "this" {
+  count = var.config.pages == null ? 0 : 1
+
+  repository    = github_repository.this.name
+  build_type    = try(var.config.pages.buildType, null)
+  cname         = try(var.config.pages.cname, null)
+
+  dynamic "source" {
+    for_each = var.config.pages != null && var.config.pages.source != null ? [var.config.pages.source] : []
     content {
-      build_type = pages.value.buildType
-      cname      = pages.value.cname
-  
-      dynamic "source" {
-        for_each = pages.value.source != null ? [pages.value.source] : []
-        content {
-          branch = source.value.branch
-          path   = coalesce(source.value.path, "/")
-        }
-      }
+      branch = source.value.branch
+      path   = coalesce(source.value.path, "/")
     }
   }
 }